requirements/requirements.txt
apache-airflow-providers-snowflake==4.0.2
apache-airflow-providers-mysql==4.0.0
apache-airflow-providers-docker
requests
python-dotenv
pytest
google-search-results
ipykernel
boto3
pytz
dbt-snowflake
lxml==4.9.1
Scrapy
scrapeops-scrapy-proxy-sdk
snowflake-connector-python==2.8.0
astronomer-cosmos[dbt-snowflake]